+++ solrorm/__init__.py
@@ -21,7 +21,7 @@ configuration by handing a :class:`Settings` instance to :class:`SolrORM`.
 """
 
 __author__ = "Valentin Grouès"
-__version__ = "0.2.0"
+__version__ = "0.3.0"
 
 from .config import Settings
 from .entity import SolrEntity